Changed entites id from int to long. New entites for tables introduced for SurveyBuilder. Renamed UserManagementServices to UserManagementService.

+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/commons/CommonConstants.java
@@ -19,12 +19,12 @@ public interface CommonConstants {
 	 * Default constants
 	 */
 	// default company id for application. if any entity is linked to this id, then its an orphan entity
-	public static final int DEFAULT_COMPANY_ID = 1;
+	public static final long DEFAULT_COMPANY_ID = 1;
 	public static final String DEFAULT_BRANCH_NAME = "Defult Branch";
 	public static final String DEFAULT_REGION_NAME = "Defult Region";
-	public static final int DEFAULT_REGION_ID = 0;
-	public static final int DEFAULT_BRANCH_ID = 0;
-	public static final int DEFAULT_AGENT_ID = 0;
+	public static final long DEFAULT_REGION_ID = 0;
+	public static final long DEFAULT_BRANCH_ID = 0;
+	public static final long DEFAULT_AGENT_ID = 0;
 	public static final String DEFAULT_SOURCE_APPLICATION = "AP";
 
 	/**
